# hotel-backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  hotel-backend_internal_models.CreateGuestRequest:
    properties:
      address:
        type: string
      email:
        type: string
      first_name:
        type: string
      id_card_or_passport:
        type: string
      last_name:
        type: string
      phone:
        type: string
    required:
    - email
    - first_name
    - id_card_or_passport
    - last_name
    - phone
    type: object
...
  hotel-backend_internal_models.UpdateGuestRequest:
    properties:
      address:
        type: string
      email:
        type: string
      first_name:
        type: string
      id_card_or_passport:
        type: string
      last_name:
        type: string
      phone:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@hotel-system.com
    name: API Support
  description: This is the API documentation for the Hotel Management System.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Hotel Management System API
  version: "1.0"
paths:
...
  /guests:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all registered guests with optional search by
        name, email, or phone
      parameters:
      - description: Search query (name, email, phone)
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Guests retrieved
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all guests
      tags:
      - Guests
    post:
      consumes:
      - application/json
      description: Add a new guest profile to the system (Admin and Receptionist)
      parameters:
      - description: Guest data
        in: body
        name: guest
        required: true
        schema:
          $ref: '#/definitions/hotel-backend_internal_models.CreateGuestRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Guest registered successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Register a new guest
      tags:
      - Guests
  /guests/{id}:
    delete:
      consumes:
      - application/json
      description: Remove a guest record from the system (Admin only)
      parameters:
      - description: Guest ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Guest deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid guest ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - Admin only
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to delete guest
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete guest
      tags:
      - Guests
    get:
      consumes:
      - application/json
      description: Retrieve details of a specific guest by ID
      parameters:
      - description: Guest ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Guest retrieved
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid guest ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Guest not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get guest by ID
      tags:
      - Guests
    put:
      consumes:
      - application/json
      description: Modify an existing guest's contact or identification details
      parameters:
      - description: Guest ID
        in: path
        name: id
        required: true
        type: integer
      - description: Guest update data
        in: body
        name: guest
        required: true
        schema:
          $ref: '#/definitions/hotel-backend_internal_models.UpdateGuestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Guest updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Guest not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update guest
      tags:
      - Guests
  /reports/occupancy:
    get:
      consumes:
      - application/json
      description: Calculate current hotel occupancy rate and room status distribution
        (Admin and Receptionist)
      produces:
      - application/json
      responses:
        "200":
          description: Occupancy report retrieved
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to calculate occupancy report
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get occupancy report
      tags:
      - Reports
  /reports/revenue:
    get:
      consumes:
      - application/json
      description: Calculate revenue metrics by room type, payment method, and summary
        within an optional date range (Admin only)
      parameters:
      - description: From date (YYYY-MM-DD)
        in: query
        name: from
        type: string
      - description: To date (YYYY-MM-DD)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Revenue report retrieved
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - Admin only
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to calculate revenue report
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get revenue report
      tags:
      - Reports
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: 'Enter your JWT token as: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"